# openapi.yaml
openapi: 3.0.3
info:
  title: Stats API
  version: 1.0.0
  description: Endpoints exposed by src/app.ts (Koa).
servers:
  - url: http://localhost:8000/v1
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: token
    JwtAuth:
      type: apiKey
      in: header
      name: jwt-token
    PushAuth:
      type: apiKey
      in: header
      name: push-token
  schemas:
...
    LiveStreamSummary:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        app: { type: string }
        viewers: { type: integer }
        duration: { type: number }
        bitrate: { type: number }
        lastBitrate: { type: number }
        startTime: { type: string, format: date-time }
        protocol: { type: string }
        userName: { type: string, nullable: true }
        origin: { type: string }
        urls:
          type: object
          nullable: true
          properties:
            web: { type: string }
            edge: { type: string }
...
paths:
...
  /channels/{channel}:
    get:
      summary: List live streams in a channel
      parameters:
        - name: channel
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  streams:
                    type: array
                    items:
                      $ref: '#/components/schemas/LiveStreamSummary'
...
  /admin/channels:
    get:
      summary: List channels (admin only)
      security:
        - TokenAuth: []
        - JwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  channels:
                    type: array
                    items: { type: object }
  /admin/channels/{id}:
    put:
      summary: Update channel type (admin only)
      security:
        - TokenAuth: []
        - JwtAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [PUBLIC, PRIVATE]
      responses:
        '201':
          description: Channel updated
...